requirements.txt
# Core Python dependencies for Research Assistant Agent
# Python version: 3.9 or higher

# Web requests and HTML parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
numpy==1.24.3

# API and search
newsapi==0.1.1

# NLP and text processing
nltk==3.8.1

# Date/time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Optional: For better CLI output
colorama==0.4.6

# Optional: For markdown formatting
markdown==3.4.1

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0
